server/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pypdf
sentence-transformers
